Notifications_management/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
authlib==1.3.1
itsdangerous==2.1.2
pytest==8.4.2
requests==2.32.5
aiosmtplib==4.0.2
twilio==9.8.0